增加了权限管理！！！

+++ trunk/串口服务器/软件/nc/inc/conf.h
@@ -59,7 +59,7 @@ typedef struct tagCONFIG_COM
 {
     DWORD id;
     DWORD bps;
-    float sbit;
+    DWORD sbit;
     BYTE  dbit;
     BYTE  chk;
     BYTE  bctrl;
